LU-8066 obd: migrate several subsystems to debugfs
move status files from procfs to debugfs

change type of
 * obd->obd_proc_entry
 * obd->obd_svc_procroot
 * lov->lov_pool_proc_entry
 * obd_type->typ_procroot
 * pool_desc->pool_proc_entry

Linux-commit:  61e87ab0f8b24300f10ea345ebf382d9f93c5bc2

Port lprocfs_alloc_md_stats which is missing upstream
to debugfs. Add a special LDEBUGFS_FOPS_WR_ONLY since
often in this case i_private is not set so it attempts
to call PDE_DATA(inode) which causes it to crash. Make
lprocfs_obd_setup select either debugfs or procfs but
not both.

Handle the special symlinks needed for both debugfs
and sysfs with the server case. For lod we need to
create "lov" and osp we create "osc" for both sysfs
and debugfs. Handle the complex case of when a node
is both a server and client. For debugfs we can take
advantage of d_lookup() and for sysfs kset_find_obj()
to avoid special access to struct obd_type. This also
places the burden on the server lod/osp modules instead
of the client lov/osc modules.

Change-Id: I87090859db4da2300ab9e2aa3c23cb3773276103
Signed-off-by: Dmitry Eremin <dmiter4ever@gmail.com>
Signed-off-by: Oleg Drokin <green@linuxhacker.ru>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: James Simmons <uja.ornl@yahoo.com>